backend/requirements.txt
flask
flask-cors
requests
requests-oauthlib
python-dotenv
sqlalchemy